refactored application to be a multi-module project
