+++ consumer_app/Dockerfile
