+++ consumer_app/src/main/resources/application.properties
@@ -1,0 +1,13 @@
+#Server config
+spring.application.name=consumer-app
+server.port=8082
+
+#Kafka config
+spring.kafka.bootstrap-servers=kafka:9092
+spring.kafka.consumer.group-id=order-group
+spring.kafka.consumer.auto-offset-reset=earliest
+spring.kafka.consumer.key-deserializer=org.apache.kafka.common.serialization.StringDeserializer
+spring.kafka.consumer.value-deserializer=org.apache.kafka.common.serialization.StringDeserializer
+
+spring.kafka.listener.missing-topics-fatal=false
+
