+++ docker-compose.yml
@@ -22,13 +22,13 @@ services:
 
   producer:
     build:
-      context: src/main/java/taras/practice/producer_app
+      context: ./producer_app
     depends_on:
       - kafka
 
   consumer:
     build:
-      context: src/main/java/taras/practice/consumer_app
+      context: ./consumer_app
     depends_on:
       - kafka
     deploy:
