+++ producer_app/src/main/java/taras/producer/util/MessageEntity.java
@@ -1,0 +1,4 @@
+package taras.producer.util;
+
+public record MessageEntity(String message) {
+}
