+++ producer_app/src/main/resources/application.properties
@@ -1,0 +1,15 @@
+#Server config
+spring.application.name=producer-app
+server.port=8081
+
+#Kafka config
+spring.kafka.bootstrap-servers=kafka:9092
+spring.kafka.producer.key-serializer=org.apache.kafka.common.serialization.StringSerializer
+spring.kafka.producer.value-serializer=org.apache.kafka.common.serialization.StringSerializer
+
+#Optional topic config
+#app.kafka.topic=orders
+
+
+
+
