--- src/main/java/taras/practice/KafkaOrderSystemApplication.java
@@ -1,13 +1,0 @@
-package taras.practice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class KafkaOrderSystemApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(KafkaOrderSystemApplication.class, args);
-	}
-
-}
