--- src/main/java/taras/practice/consumer_app/src/main/java/taras/consumer/ConsumerApp.java
@@ -1,11 +1,0 @@
-package taras.practice.consumer_app.src.main.java.taras.consumer;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ConsumerApp {
-    public static void main(String[] args) {
-        SpringApplication.run(ConsumerApp.class, args);
-    }
-}
