--- src/main/java/taras/practice/consumer_app/src/main/resources/application.properties
@@ -1,12 +1,0 @@
-# ????? ????? ??????????
-spring.kafka.consumer.group-id=consumer-group
-
-# ???????????? ?????/????????
-spring.kafka.consumer.key-deserializer=org.apache.kafka.common.serialization.StringDeserializer
-spring.kafka.consumer.value-deserializer=org.apache.kafka.common.serialization.StringDeserializer
-
-# ?????? Kafka ??????? (?? ??'? ??????? ? docker-compose)
-spring.kafka.bootstrap-servers=kafka:9092
-
-# (???????????) ???? ?? ????????????? Spring Boot 3.1+, ???? ???????????
-spring.kafka.listener.missing-topics-fatal=false
