--- src/main/java/taras/practice/producer_app/src/main/java/taras/producer/ProducerApp.java
@@ -1,11 +1,0 @@
-package taras.practice.producer_app.src.main.java.taras.producer;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ProducerApp {
-    public static void main(String[] args) {
-        SpringApplication.run(ProducerApp.class, args);
-    }
-}
