--- src/main/java/taras/practice/producer_app/src/main/java/taras/producer/util/MessageEntity.java
@@ -1,4 +1,0 @@
-package taras.practice.producer_app.src.main.java.taras.producer.util;
-
-public record MessageEntity(String message) {
-}
