--- src/main/java/taras/practice/producer_app/src/main/resources/application.properties
@@ -1,6 +1,0 @@
-# ?????? Kafka ??????? (?? ???? ?? ? consumer)
-spring.kafka.bootstrap-servers=kafka:9092
-
-# ???????????? ??? ???????????? ???????????
-spring.kafka.producer.key-serializer=org.apache.kafka.common.serialization.StringSerializer
-spring.kafka.producer.value-serializer=org.apache.kafka.common.serialization.StringSerializer
