--- src/test/java/taras/practice/KafkaOrderSystemApplicationTests.java
@@ -1,13 +1,0 @@
-package taras.practice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class KafkaOrderSystemApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
